+++ functions/i/[id].js
@@ -1,0 +1,1 @@
+export { onRequest } from "../utils/file.js";