+++ functions/i/_middleware.js
@@ -1,0 +1,3 @@
+import { errorHandling } from "../utils/middleware.js";
+
+export const onRequest = [errorHandling];